minor bug

+++ src/services/apiFriends.js
@@ -13,23 +13,29 @@ export async function getFriends() {
 //https://vhbqdebzwitgkiedgkoc.supabase.co/storage/v1/object/public/avatars/avatar-1447bed6-777a-4572-bfac-8aea0b1eb6f1-0.3620885902466795
 
 export async function createFriend (friend) {
+
+    const isAvatar = friend.avatar?true:false;
+     
     //creating friend and avatar as url in friends table.
-    const avatarUrl = `${Math.random()}-${friend.avatar.name}`.replace("/","");
+    const avatarUrl = isAvatar?`${Math.random()}-${friend.avatar.name}`.replace("/",""):null;
     
-    const avatarPath = `${supabaseUrl}/storage/v1/object/public/avatars/${avatarUrl}`;
+    const avatarPath = isAvatar?`${supabaseUrl}/storage/v1/object/public/avatars/${avatarUrl}`: null;
 
         const {data,error } = await supabase.from("friends").insert([{...friend, avatar:avatarPath}]).select();
 
     if(error) {throw new Error (error.message)}
 
-    // storing avatar image in avatar bucket
-    const{error:error2} = await supabase.storage.from("avatars").upload(`${avatarUrl}`, friend.avatar);
+    if(isAvatar) { 
+         // storing avatar image in avatar bucket
+        const{error:error2} = await supabase.storage.from("avatars").upload(`${avatarUrl}`, friend.avatar);
+     if(error2) {
+            //3. If image is not uploaded we would want to delete the friend itself which we can get hold of via data in 21 line.
 
-    //3. If image is not uploaded we would want to delete the friend itself which we can get hold of via data in 21 line.
-    if(error2) {
         await supabase.from("friends").delete().eq("id",data.id);
         throw new Error ("Avatar could not be uploaded and friend has not been created.");
-    } 
+    } }
+  
+   
 }
